Move reference to daemon/internal/refstore
