+++ daemon/daemon.go
@@ -46,6 +46,7 @@ import (
 	libcontainerdtypes "github.com/docker/docker/daemon/internal/libcontainerd/types"
 	"github.com/docker/docker/daemon/internal/metrics"
 	pluginexec "github.com/docker/docker/daemon/internal/plugin/executor/containerd"
+	refstore "github.com/docker/docker/daemon/internal/refstore"
 	"github.com/docker/docker/daemon/libnetwork"
 	"github.com/docker/docker/daemon/libnetwork/cluster"
 	nwconfig "github.com/docker/docker/daemon/libnetwork/config"
@@ -63,7 +64,6 @@ import (
 	"github.com/docker/docker/pkg/idtools"
 	"github.com/docker/docker/pkg/plugingetter"
 	"github.com/docker/docker/pkg/sysinfo"
-	refstore "github.com/docker/docker/reference"
 	"github.com/docker/docker/registry"
 	"github.com/moby/buildkit/util/grpcerrors"
 	"github.com/moby/buildkit/util/tracing"
