+++ daemon/internal/builder-next/adapters/containerimage/pull.go
@@ -30,8 +30,8 @@ import (
 	"github.com/docker/docker/daemon/internal/distribution/xfer"
 	"github.com/docker/docker/daemon/internal/image"
 	"github.com/docker/docker/daemon/internal/layer"
+	refstore "github.com/docker/docker/daemon/internal/refstore"
 	pkgprogress "github.com/docker/docker/pkg/progress"
-	refstore "github.com/docker/docker/reference"
 	"github.com/moby/buildkit/cache"
 	"github.com/moby/buildkit/client"
 	"github.com/moby/buildkit/client/llb/sourceresolver"
