+++ daemon/internal/distribution/config.go
@@ -13,8 +13,8 @@ import (
 	"github.com/docker/docker/daemon/internal/distribution/xfer"
 	"github.com/docker/docker/daemon/internal/image"
 	"github.com/docker/docker/daemon/internal/layer"
+	refstore "github.com/docker/docker/daemon/internal/refstore"
 	"github.com/docker/docker/pkg/progress"
-	refstore "github.com/docker/docker/reference"
 	registrypkg "github.com/docker/docker/registry"
 	"github.com/moby/moby/api/types/events"
 	"github.com/moby/moby/api/types/registry"
