+++ daemon/internal/distribution/push_v2_test.go
@@ -12,8 +12,8 @@ import (
 	"github.com/docker/distribution/registry/api/errcode"
 	"github.com/docker/docker/daemon/internal/distribution/metadata"
 	"github.com/docker/docker/daemon/internal/layer"
+	refstore "github.com/docker/docker/daemon/internal/refstore"
 	"github.com/docker/docker/pkg/progress"
-	refstore "github.com/docker/docker/reference"
 	registrypkg "github.com/docker/docker/registry"
 	"github.com/moby/moby/api/types/registry"
 	"github.com/opencontainers/go-digest"
