+++ daemon/internal/image/tarexport/tarexport.go
@@ -7,7 +7,7 @@ import (
 	"github.com/docker/distribution"
 	"github.com/docker/docker/daemon/internal/image"
 	"github.com/docker/docker/daemon/internal/layer"
-	refstore "github.com/docker/docker/reference"
+	refstore "github.com/docker/docker/daemon/internal/refstore"
 	"github.com/moby/moby/api/types/events"
 	ocispec "github.com/opencontainers/image-spec/specs-go/v1"
 )
