+++ daemon/internal/refstore/errors.go
